+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [Unreleased]
+## 1.0.1 (2026-05-26)
 - Added basic define macro support
 - Removed unsupported matrix constructors
 - Added sampler uniform hint completions